+++ include/symbolic/pddl.h
@@ -82,12 +82,26 @@ class Pddl {
    * derived predicates.
    *
    * @param state Current state.
-   * @param action Action call in the form of `"action(obj_a, obj_b)"`.
+   * @param action_call Action call in the form of `"action(obj_a, obj_b)"`.
    * @returns Next state.
    *
    * @seepython{symbolic.Pddl,next_state}
    */
   State NextState(const State& state, const std::string& action_call) const;
+
+  /**
+   * Execute a sequence of actions from the given state.
+   *
+   * The action preconditionss are not checked. The resulting state includes
+   * derived predicates.
+   *
+   * @param state Current state.
+   * @param action_call Action calls in the form of `"action(obj_a, obj_b)"`.
+   * @returns Final state.
+   *
+   * @seepython{symbolic.Pddl.execute}
+   */
+  State ApplyActions(const State& state, const std::vector<std::string>& action_calls) const;
 
   /**
    * Apply the derived predicates to the given state.
@@ -185,6 +199,9 @@ class Pddl {
   std::vector<std::string> ListValidActions(
       const std::set<std::string>& state) const;
 
+  void AddObject(const std::string& name, const std::string& type);
+  void RemoveObject(const std::string& name);
+
   const VAL::analysis* symbol() const { return analysis_.get(); }
 
   /**
@@ -210,6 +227,7 @@ class Pddl {
 
   const ObjectTypeMap& object_map() const { return object_map_; }
 
+  const std::vector<Object>& constants() const { return constants_; }
   const std::vector<Object>& objects() const { return objects_; }
 
   const std::vector<Action>& actions() const { return actions_; }
@@ -236,6 +254,7 @@ class Pddl {
   std::string domain_pddl_;
   std::string problem_pddl_;
 
+  std::vector<Object> constants_;
   std::vector<Object> objects_;
   ObjectTypeMap object_map_;