+++ symbolic/__init__.py
@@ -1,6 +1,6 @@
 from .pysymbolic import *
-from .problem import Problem, _P, _and, _or, _not, _forall, _exists, _when, parse_head, parse_args
+from .problem import Problem, _P, _and, _or, _not, _forall, _exists, _when, parse_proposition, parse_head, parse_args
 
-__version__ = "0.1.2"
+__version__ = "0.3.7"
 
 # __doc__ = pysymbolic.__doc__