+++ CHANGELOG.md
@@ -2,6 +2,47 @@
 
 <!--next-version-placeholder-->
 
+## v0.3.7 (2022-07-13)
+### Fix
+* Allow parsing of longer object lists ([`08c95a5`](https://github.com/tmigimatsu/symbolic/commit/08c95a51563d3442cb165428f3030d30fd83adaa))
+
+## v0.3.6 (2022-07-13)
+### Fix
+* Switch file names on pddl parsing ([`476b671`](https://github.com/tmigimatsu/symbolic/commit/476b6718307e878e69d8ba5891fad73c3aa27352))
+
+## v0.3.5 (2022-07-13)
+### Fix
+* Print full pddl report ([`1a22daf`](https://github.com/tmigimatsu/symbolic/commit/1a22daf6f7f01b29edbaf6b239eb0948c3e82017))
+* Print pddl problem parsing errors ([`60ae847`](https://github.com/tmigimatsu/symbolic/commit/60ae847a6ab53323bbc7689c894ffade8ac5a0db))
+
+## v0.3.4 (2022-07-13)
+### Fix
+* Print more parsing information ([`0ad24b4`](https://github.com/tmigimatsu/symbolic/commit/0ad24b4b6f75e3f38cbc103c9780e97b49f10981))
+
+## v0.3.3 (2022-07-13)
+### Fix
+* Support empty constants and types ([`04b360c`](https://github.com/tmigimatsu/symbolic/commit/04b360c3773a2aabbc47eb6b19ed7da1c5e9f941))
+
+## v0.3.2 (2022-07-08)
+### Fix
+* Add action parsing python binding ([`5e0e61c`](https://github.com/tmigimatsu/symbolic/commit/5e0e61c72a53f4a5a24262bfe3a266e4cd2b2a97))
+
+## v0.3.1 (2022-07-01)
+### Fix
+* Pypi long description ([`834499d`](https://github.com/tmigimatsu/symbolic/commit/834499db88b5e8675968a9d80647a25b972ef7fa))
+* Foreall, exists formula serialization ([`b99271e`](https://github.com/tmigimatsu/symbolic/commit/b99271ebc6f23ad017289d049fb35425c8a9c92b))
+
+## v0.3.0 (2022-05-04)
+### Feature
+* Support dynamically adding objects and serializing pddl ([`0d5e50f`](https://github.com/tmigimatsu/symbolic/commit/0d5e50f24a719de4db7f614b89bec3101e332559))
+
+## v0.2.0 (2022-04-26)
+### Feature
+* Implement fast method to apply sequence of actions ([`b7a4a68`](https://github.com/tmigimatsu/symbolic/commit/b7a4a68f406782a808b50d5875d4a5266f47146d))
+
+### Fix
+* Build only cpython wheels ([`20ac4ac`](https://github.com/tmigimatsu/symbolic/commit/20ac4ace704584a9765345e5cb08b2750623a793))
+
 ## v0.1.2 (2022-01-15)
 ### Fix
 * Do not upload release ([`b5e172b`](https://github.com/tmigimatsu/symbolic/commit/b5e172b572b2d3fb4410c378315e5228332ef8d6))